AN-1250 Add Apk Tags Beta/Alpha (cherry picked from commit e3f78a8)
