+++ dataprovider/src/main/java/cm/aptoide/pt/dataprovider/ws/v7/BaseBodyWithApp.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 /**
  * Created by diogoloureiro on 12/09/16.
  */
-@EqualsAndHashCode(callSuper = true) public class BaseBodyWithApp extends BaseBody {
+@EqualsAndHashCode(callSuper = true) public class BaseBodyWithApp extends BaseBodyWithAlphaBetaKey {
   @Getter @Setter private String storeUser;
   @Getter @Setter private String storePassSha1;
 }
